api/requirements.txt
fastapi
uvicorn
mlflow
boto3
pydantic
python-dotenv
